requirements.txt
Jinja2
Werkzeug
pytz
isoweek
odict
pyyaml
lxml
beautifulsoup4
numpy
matplotlib
git+git://github.com/mwclient/mwclient.git
git+git://github.com/danmichaelo/mwtemplates.git
git+git://github.com/danmichaelo/mwtextextractor.git
datetime
rollbar
flipflop
flask
requests
mysql-connector==2.1.6
psutil
